Update Basic_calculator to use float for user input and add interest_calculator program

+++ python/Py4e(Book)/Practice_1-6/Chapter_2/Basic_calculator.py
@@ -1,6 +1,6 @@
-num1 = int(input("Enter the first number: "))
+num1 = float(input("Enter the first number: "))
 operator = input("Enter the operator(+, -, /, *): ")
-num2 = int(input("Enter the second number: "))
+num2 = float(input("Enter the second number: "))
 
 if operator == "+":
     result = num1 + num2
